+++ upload/admin/language/en-gb/extension/shipping/omniva_m.php
@@ -4,6 +4,7 @@
 $_['omniva_m_menu_head'] = 'Omniva';
 $_['omniva_m_menu_manifest'] = 'Manifest';
 $_['omniva_m_menu_settings'] = 'Settings';
+$_['omniva_m_menu_called_couriers'] = 'Active Courier Calls';
 
 // Heading - without prefix as thats what opencart expects
 $_['heading_title'] = 'Omniva';
@@ -80,7 +81,7 @@ $_['omniva_m_help_order_status_registered'] = 'What order status to set after su
 $_['omniva_m_label_order_status_error'] = 'Register error status';
 $_['omniva_m_help_order_status_error'] = 'What order status to set after failed label registration.';
 $_['omniva_m_label_disable_cart_weight_check'] = 'Disable cart weight check for terminals';
-$_['omniva_m_help_disable_cart_weight_check'] = 'If set to YES disables total cart weight check (max 30kg) for terminals';
+$_['omniva_m_help_disable_cart_weight_check'] = 'If set to YES disables total cart weight check (max 25kg) for terminals';
 $_['omniva_m_label_use_simple_terminal_check'] = 'Use simple terminal fit check';
 $_['omniva_m_help_use_simple_terminal_check'] = 'Checks if cart fits terminal box dimensions W38 - H39 - L64.';